Add ConsensusStateBytesQuerier. Move QueryAndDecodeClientStateVia
