+++ crates/relayer/relayer-components/src/chain/impls/delegate/queries/client_state.rs
@@ -1,16 +1,9 @@
 use alloc::vec::Vec;
 use core::marker::PhantomData;
 
-use cgp_core::{Async, CanRaiseError, DelegateComponent, HasErrorType};
-use hermes_encoding_components::traits::decoder::CanDecode;
-use hermes_encoding_components::traits::encoded::HasEncodedType;
-use hermes_encoding_components::traits::has_encoding::HasDefaultEncoding;
-use hermes_encoding_components::types::via::Via;
+use cgp_core::{DelegateComponent, HasErrorType};
 
-use crate::chain::traits::queries::client_state::{
-    AllClientStatesQuerier, CanQueryAllClientStatesBytes, CanQueryClientStateBytes,
-    ClientStateQuerier,
-};
+use crate::chain::traits::queries::client_state::{AllClientStatesQuerier, ClientStateQuerier};
 use crate::chain::traits::types::client_state::HasClientStateType;
 use crate::chain::traits::types::ibc::HasIbcChainTypes;
 
@@ -46,60 +39,5 @@ where
         height: &Chain::Height,
     ) -> Result<Vec<(Chain::ClientId, Counterparty::ClientState)>, Chain::Error> {
         Delegate::query_all_client_states(chain, height).await
-    }
-}
-
-pub struct QueryAndDecodeClientStateVia<Wrapper>(pub PhantomData<Wrapper>);
-
-impl<Chain, Counterparty, Encoding, Wrapper> ClientStateQuerier<Chain, Counterparty>
-    for QueryAndDecodeClientStateVia<Wrapper>
-where
-    Chain: CanQueryClientStateBytes<Counterparty> + CanRaiseError<Encoding::Error>,
-    Counterparty: HasClientStateType<Chain> + HasDefaultEncoding<Encoding = Encoding>,
-    Encoding:
-        HasEncodedType<Encoded = Vec<u8>> + CanDecode<Via<Wrapper, Counterparty::ClientState>>,
-    Wrapper: Async,
-{
-    async fn query_client_state(
-        chain: &Chain,
-        client_id: &Chain::ClientId,
-        height: &Chain::Height,
-    ) -> Result<Counterparty::ClientState, Chain::Error> {
-        let client_state_bytes = chain.query_client_state_bytes(client_id, height).await?;
-
-        let client_state = Counterparty::default_encoding()
-            .decode(&client_state_bytes)
-            .map_err(Chain::raise_error)?;
-
-        Ok(client_state.value)
-    }
-}
-
-impl<Chain, Counterparty, Encoding, Wrapper> AllClientStatesQuerier<Chain, Counterparty>
-    for QueryAndDecodeClientStateVia<Wrapper>
-where
-    Chain: CanQueryAllClientStatesBytes<Counterparty>,
-    Counterparty: HasClientStateType<Chain> + HasDefaultEncoding<Encoding = Encoding>,
-    Encoding:
-        HasEncodedType<Encoded = Vec<u8>> + CanDecode<Via<Wrapper, Counterparty::ClientState>>,
-    Wrapper: Async,
-{
-    async fn query_all_client_states(
-        chain: &Chain,
-        height: &Chain::Height,
-    ) -> Result<Vec<(Chain::ClientId, Counterparty::ClientState)>, Chain::Error> {
-        let encoding = Counterparty::default_encoding();
-        let raw_entries = chain.query_all_client_states_bytes(height).await?;
-
-        let entries = raw_entries
-            .into_iter()
-            .filter_map(|(client_id, client_state_bytes)| {
-                let client_state = encoding.decode(&client_state_bytes).ok()?;
-
-                Some((client_id, client_state.value))
-            })
-            .collect();
-
-        Ok(entries)
     }
 }
