+++ crates/relayer/relayer-components/src/chain/impls/mod.rs
@@ -1,2 +1,3 @@
 pub mod delegate;
+pub mod queries;
 pub mod wait_chain_reach_height;
